Create Deploy2.dolfined.yaml
This file defines a Kubernetes Pod named "dolfined" using the CentOS image. The container mounts a local directory from the host (/home/ubuntu/dolfined) to the path /dats inside the container using a hostPath volume. The container runs a simple script that appends the current date and time to a file every 5 seconds. Additionally, a nodeSelector is used to schedule the Pod specifically on node1.