Create pod3.yaml
This file defines a PersistentVolumeClaim (PVC) in Kubernetes named dolfined-pvc. It requests 1Gi of storage and specifies the ReadWriteMany (RWX) access mode, which allows multiple nodes to read from and write to the volume simultaneously. The actual volume must support this access mode (e.g., NFS or EFS). The PVC will bind to a matching PersistentVolume (PV) based on storage size and access capabilities