Create Create StorageClass for MySQL using AWS EBS CSI Driver
This manifest defines a Kubernetes StorageClass named mysql-sc that leverages the ebs.csi.aws.com provisioner to dynamically provision Amazon EBS volumes. The volumeBindingMode is set to WaitForFirstConsumer, which means the volume will only be provisioned when a Pod using the PVC is scheduled, ensuring it is created in the correct Availability Zone. This setup is typically used for deploying stateful workloads such as MySQL in AWS-backed Kubernetes clusters.